feat(view): ViewController를 정적 리소스 응답 전용으로 리팩터링 및 Fallback 핸들러로 추가
